Implement initial integration test
The test will basically simulate an incoming pull request event
and synchronize it with a dummy repository. The incoming pull request
event is taken from the GitHub documentation.
https://developer.github.com/v3/activity/events/types/#pullrequestevent

This adds the Symfony PHPUnit bridge and an assortment of other 
packages to support testing. 

Safe is used to handle functions which normally return false when 
errors occur. Now they throw exceptions which can be handled more
elegantly and explicitly.