Implemented logging of requests/responses
Use the http-client framework to use a logger plugin and client to
wrap the core http client.

We could also use Guzzle to achieve this but the premiere Symfony
bundle which also has support for logging uses deprecated code which
makes the logs noisy. This way we avoid using that bundle and also
decouple from Guzzle itself.

A sneak peak of the Platform.sh Client which we will probably need
later on also shows that is does not support a preconfigured 
Guzzle client with logging so that will not help us anyway.