Implement support for GitHub webhooks
We use a fork of a third-party bundle with Symfony 4 support to
achieve this. Perhaps this could be achieved in a better way but this
does the job.

We store the secret in the environment variables.

Our handler only supports pull request events but the initial setup
of a webhook will send a ping event. Keep this check in the
controller and return quickly.