Rework event handling to use messages
Webhooks should respond in less than 10 seconds. We cannot expect
our deployment to Platform.sh to complete within that timeframe so
we prepare to complete our work asynchronously. Symfony Messager 
provides a way for us to do so.

For now we primarily install the package and use a message bus to
dispatch GitHub events to a handler which performs the actual
work.