Use a private key to use when pushing to repo
This is needed for pushing to Platform.sh as they do not support any
other from of authentication.

To be somewhat cloud compatible we use an environment variable for
holding the private key. GitWrapper needs it to be in a file so we
create a file containing the key and use that before running any
commands.

Update environment variables to have the GIT_ prefix for both key and
repository.