Reworked key handling
Git can be prickly about malformed keys resulting in access denied
errors even though the private key file is actually malformed.

Malformed keys can be the result of wrongly handled environment
variables with our without line breaks and how they are processed.

Consequently we introduce a new class for handling private keys
which will do its best to identify invalid keys and do some
processing to make them valid. The most prominent example of this
is to add a valid header and footer with line breaks when the key
itself is provided as a single string.

Add test accordingly.