Use loggy for logging
We really need a logger which will output the context. The Symfony
default logger does not do this.

Monolog has caused problems getting our workers to pick up events
when using Docker so we try something else.

Logge seems to fit the bill: Small, no dependencies and outputs to
JSON which CloudWatch likes.