Retry Platform.sh environment url retrieval
Sometimes an environment may not have any urls available despite 
being active and having no pending activities.

We try to fix this by making it clear that the environment manager
should return an url (string) and otherwise throw an exception. This
allows to use backoff to retry url retrieval if the environment has
no urls.