Rework activation and notification after synchronization
The current approach with two handlers listening for the same event
causes problems: The notification risks waiting and failing because
the activation has not been executed.

Consequently we change this from parallel to sequential steps by
introducing a new event type and only try to notify when we know that
the environment is ready.